+++ src/key.py
@@ -1,1 +1,1 @@
-key = "AIzaSyApqXMMJ0S-hQxDDFNnnn54GFmt-Z1KpRk"
+key = ""
