+++ src/output/summary/Balkan trip (1).json
@@ -1,29 +1,8 @@
 {
-  "summary": "",
-  "keywords": [
-    {
-      "keyword": "keckemet serbia",
-      "score": 0.5154
-    },
-    {
-      "keyword": "mostar bosnia",
-      "score": 0.5226
-    },
-    {
-      "keyword": "croatia nights",
-      "score": 0.57
-    },
-    {
-      "keyword": "albania night",
-      "score": 0.5886
-    },
-    {
-      "keyword": "balkans route",
-      "score": 0.6241
-    }
-  ],
-  "topics": [],
-  "token_count": null,
-  "mode": "keybert",
+  "summary": "This document outlines a planned 2026 Balkan travel itinerary, offering two transportation options: driving from Northern Europe or using cross-border buses. The route covers major cities in Serbia, Bosnia, Croatia, Montenegro, Albania, Greece, Bulgaria, and Romania, including detailed schedules, accommodation recommendations in Shkoder, and logistical tips for cross-border travel.",
+  "keywords": "Balkans, Road trip, Itinerary, Transportation, Tourism",
+  "topics": "Travel planning, Western Balkans, Transportation logistics",
+  "token_count": 2036,
+  "mode": "genai",
   "source_file": "C:\\Users\\Pihla\\Downloads\\Balkan trip (1).txt"
 }
