+++ src/output/summary/peerj-cs-967 (1).json
@@ -1,0 +1,29 @@
+{
+  "summary": "",
+  "keywords": [
+    {
+      "keyword": "keywords comparable",
+      "score": 0.5452
+    },
+    {
+      "keyword": "keywords proposed",
+      "score": 0.5459
+    },
+    {
+      "keyword": "keyword extraction",
+      "score": 0.5513
+    },
+    {
+      "keyword": "automatic keyword",
+      "score": 0.5548
+    },
+    {
+      "keyword": "keywords retrieved",
+      "score": 0.5597
+    }
+  ],
+  "topics": [],
+  "token_count": null,
+  "mode": "keybert",
+  "source_file": "C:\\Users\\Pihla\\Downloads\\peerj-cs-967 (1).pdf"
+}
