+++ src/send_prompt_online.py
@@ -3,42 +3,25 @@ import os
 from google import genai
 import json
 import re
+import state
 
 try:
     from keybert_analyzer import get_keybert_keywords
 except ImportError:
     from .keybert_analyzer import get_keybert_keywords
 
-ai_model = "gemini-3-flash-preview"
-
-# initial values only, later will be assigned in a function based
-# on the file word count
-max_summary = 60
-max_keywords = 5
-max_topics = 3
-
-json_format = {
-    "summary": "summary here",
-    "keywords": "keywords here",
-    "topics": "topics here",
-}
-
-def get_output(message):
-    api_key = os.environs.get("GEMINI_API_KEY")
-    client = genai.Client(api_key=api_key)
 
 # send prompt to the AI
 def get_output(message):
-    api_key = os.environ.get("GEMINI_API_KEY")
-    client = genai.Client(api_key=api_key)
+    client = genai.Client(api_key=state.API_KEY)
     response = client.models.generate_content(
-        model=ai_model,
+        model=state.AI_MODEL,
         contents="Your task is to create keywords, topics, "
         + "and a summary for the provided text. "
-        + f"Maximum number of words in the summary is {max_summary}. "
-        + f"Maximum number of keywords is {max_keywords}. "
-        + f"Maximum number of topics is {max_topics}. "
-        + f"Only respond with a json object of this format: {json_format}. "
+        + f"Maximum number of words in the summary is {state.MAX_SUMMARY}. "
+        + f"Maximum number of keywords is {state.MAX_KEYWORDS}. "
+        + f"Maximum number of topics is {state.MAX_TOPICS}. "
+        + f"Only respond with a json object of this format: {state.JSON_FORMAT}. "
         + f"Here are the contents that you need to read: {message}",
     )
     return response
@@ -56,7 +39,7 @@ def get_output_text(output):
 #   keybert: use KeyBERT for keywords only (no summary or topics)
 def analyze_text(message, mode="genai", top_n_keywords=None):
     if top_n_keywords is None:
-        top_n_keywords = max_keywords
+        top_n_keywords = state.MAX_KEYWORDS
 
     result = {
         "summary": "",
@@ -69,7 +52,6 @@ def analyze_text(message, mode="genai", top_n_keywords=None):
 
     if mode == "genai":
         response = get_output(message)
-        
         parsed = json.loads(get_output_text(response))
         result["summary"] = parsed.get("summary", "")
         result["ai_keywords"] = parsed.get("keywords", [])
