+++ src/user_data.json
@@ -1,0 +1,3 @@
+{
+    "api key": ""
+}
